refactor(roles): relocate assign_SAAMR_roles from properties to roles
