+++ mupt/mupr/properties.py
@@ -7,7 +7,6 @@ __author__ = "Joseph R. Laforet Jr."
 __email__ = "jola3134@colorado.edu"
 
 from .primitives import Primitive
-from .roles import PrimitiveRole
 
 
 def is_SAAMR_compliant(prim: Primitive) -> bool:
@@ -25,32 +24,3 @@ def is_SAAMR_compliant(prim: Primitive) -> bool:
    """
 
    return all(leaf.is_atom and (leaf.depth == 3) for leaf in prim.leaves)
-
-
-def assign_SAAMR_roles(prim : Primitive) -> None:
-   """Assign canonical export roles for a SAAMR-compliant hierarchy.
-
-   Parameters
-   ----------
-   prim : Primitive
-      Root Primitive of a hierarchy expected to follow SAAMR layout:
-      universe -> segment -> residue -> particle.
-
-   Raises
-   ------
-   ValueError
-      If ``prim`` is not SAAMR-compliant.
-   """
-   if not is_SAAMR_compliant(prim):
-      raise ValueError('Cannot assign SAAMR roles: hierarchy is not SAAMR-compliant')
-
-   prim.role = PrimitiveRole.UNIVERSE
-
-   for segment in prim.children:
-      segment.role = PrimitiveRole.SEGMENT
-
-      for residue in segment.children:
-         residue.role = PrimitiveRole.RESIDUE
-
-         for particle in residue.children:
-            particle.role = PrimitiveRole.PARTICLE
